[frontend] FABless delete on Overview only; Fix when FF not enabled
