+++ opencti-platform/opencti-front/src/private/components/settings/users/SettingsOrganizationUserCreation.jsx
@@ -149,7 +149,14 @@ const SettingsOrganizationUserCreation = ({
       <Add />
     </Fab>
   );
-  if (isFABReplaced) { createButton = CreateOrgUserControlledDial; }
+  if (isFABReplaced) {
+    createButton = (
+      <CreateOrgUserControlledDial
+        onOpen={() => setOpenAddUser(true)}
+        onClose={() => setOpenAddUser(false)}
+      />
+    );
+  }
 
   return (
     <>
