+++ opencti-platform/opencti-front/src/private/components/settings/users/UserEdition.tsx
@@ -5,8 +5,6 @@ import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
 import Drawer, { DrawerVariant } from '@components/common/drawer/Drawer';
 import EEChip from '@components/common/entreprise_edition/EEChip';
-import { useNavigate } from 'react-router-dom';
-import { Button, Dialog, DialogActions, DialogContent, DialogTitle } from '@mui/material';
 import UserEditionConfidence from './edition/UserEditionConfidence';
 import UserEditionOrganizationsAdmin from './edition/UserEditionOrganizationsAdmin';
 import UserEditionOverview from './edition/UserEditionOverview';
@@ -19,8 +17,6 @@ import { UserPopoverEditionQuery$data } from './__generated__/UserPopoverEdition
 import Loader from '../../../../components/Loader';
 import useHelper from '../../../../utils/hooks/useHelper';
 import UpdateUserControlledDial from '../../../../components/UpdateEntityControlledDial';
-import useApiMutation from '../../../../utils/hooks/useApiMutation';
-import Transition from '../../../../components/Transition';
 
 const UserEditionFragment = graphql`
   fragment UserEdition_user on User
@@ -95,97 +91,6 @@ const UserEditionFragment = graphql`
   }
 `;
 
-const userDeletionMutation = graphql`
-  mutation UserEditionDeletionMutation($id: ID!) {
-    userEdit(id: $id) {
-      delete
-    }
-  }
-`;
-
-interface UserDeletionDialogProps {
-  userId: string,
-  handleClose: () => void;
-}
-
-const UserDeletionDialog: FunctionComponent<UserDeletionDialogProps> = ({
-  userId,
-  handleClose,
-}) => {
-  const { t_i18n } = useFormatter();
-  const navigate = useNavigate();
-  const [deleting, setDeleting] = useState<boolean>(false);
-  const [displayDelete, setDisplayDelete] = useState<boolean>(false);
-  const deleteSuccessMessage = t_i18n('', {
-    id: '... successfully deleted',
-    values: { entity_type: t_i18n('User') },
-  });
-  const [commit] = useApiMutation(
-    userDeletionMutation,
-    undefined,
-    { successMessage: deleteSuccessMessage },
-  );
-
-  const handleOpenDelete = () => setDisplayDelete(true);
-  const handleCloseDelete = () => setDisplayDelete(false);
-  const submitDelete = () => {
-    setDeleting(true);
-    commit({
-      variables: { id: userId },
-      onCompleted: () => {
-        setDeleting(false);
-        handleClose();
-        navigate('/dashboard/settings/accesses/users');
-      },
-    });
-  };
-
-  return (
-    <>
-      <Button
-        onClick={handleOpenDelete}
-        variant='contained'
-        color='error'
-        disabled={deleting}
-        sx={{ marginTop: 2 }}
-      >
-        {t_i18n('Delete')}
-      </Button>
-      <Dialog
-        open={displayDelete}
-        PaperProps={{ elevation: 1 }}
-        keepMounted={true}
-        TransitionComponent={Transition}
-        onClose={handleCloseDelete}
-      >
-        <DialogTitle>{t_i18n('Do you want to delete this user?')}</DialogTitle>
-        <DialogContent dividers>
-          <ul>
-            <li>{t_i18n('All notifications, triggers and digests associated with the user will be deleted.')}</li>
-            <li>{t_i18n('All investigations and dashboard where the user is the only admin, will be deleted.')}</li>
-          </ul>
-          {t_i18n('If you want to keep the associated information, we recommend deactivating the user instead.')}
-        </DialogContent>
-        <DialogActions>
-          <Button
-            onClick={handleCloseDelete}
-            disabled={deleting}
-          >
-            {t_i18n('Cancel')}
-          </Button>
-          <Button
-            color="secondary"
-            onClick={submitDelete}
-            disabled={deleting}
-          >
-            {t_i18n('Delete')}
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </>
-  );
-};
-
 interface UserEditionDrawerProps {
   handleClose?: () => void;
   userRef: UserPopoverEditionQuery$data['user'];
@@ -253,7 +158,6 @@ const UserEditionDrawer: FunctionComponent<UserEditionDrawerProps> = ({
         {hasSetAccess && currentTab === 4 && (
           <UserEditionConfidence user={user} context={user.editContext} />
         )}
-        <UserDeletionDialog userId={user.id} handleClose={handleClose} />
       </>)
         : (<Loader />)}
     </Drawer>
