+++ opencti-platform/opencti-front/src/private/components/settings/users/edition/UserEditionOverview.tsx
@@ -21,6 +21,8 @@ import { isOnlyOrganizationAdmin } from '../../../../../utils/hooks/useGranted';
 import useApiMutation from '../../../../../utils/hooks/useApiMutation';
 import PasswordTextField from '../../../../../components/PasswordTextField';
 import type { Theme } from '../../../../../components/Theme';
+import useHelper from '../../../../../utils/hooks/useHelper';
+import UserDeletionDialog from '../UserDeletionDialog';
 
 export const userMutationFieldPatch = graphql`
   mutation UserEditionOverviewFieldPatchMutation(
@@ -103,6 +105,8 @@ UserEditionOverviewComponentProps
   const [commitFieldPatch] = useApiMutation(userMutationFieldPatch);
   const [commitOrganizationAdd] = useApiMutation(userMutationOrganizationAdd);
   const [commitOrganizationDelete] = useApiMutation(userMutationOrganizationDelete);
+  const { isFeatureEnable } = useHelper();
+  const isFABReplaced = isFeatureEnable('FAB_REPLACEMENT');
 
   const userIsOnlyOrganizationAdmin = isOnlyOrganizationAdmin();
   const external = user.external === true;
@@ -323,6 +327,9 @@ UserEditionOverviewComponentProps
             onFocus={handleChangeFocus}
             onChange={handleSubmitField}
           />
+          {isFABReplaced && (
+            <UserDeletionDialog userId={user.id} />
+          )}
         </Form>
       )}
     </Formik>
